chore : designer feedback
