+++ apps/website/components/atoms/task-icon/index.ts
@@ -1,0 +1,2 @@
+export * from './task-icon';
+export * from './types';
