+++ apps/website/components/atoms/task-icon/task-icon.tsx
@@ -11,19 +11,7 @@ import Box from '@mui/material/Box';
 import { SvgIcon, SvgIconProps } from '@mui/material';
 import { useMemo } from 'react';
 import { SxProps } from '@mui/material';
-
-export type TaskType =
-  | 'self_verify'
-  | 'quiz'
-  | 'token_hold'
-  | 'nft_hold'
-  | 'meeting_code'
-  | 'twitter_follow'
-  | 'twitter_retweet'
-  | 'twitter_tweet'
-  | 'github_contribute'
-  | 'github_prs'
-  | 'snapshot';
+import { TaskType } from './types';
 
 export function TaskIcon({ type, sx }: { type: TaskType; sx?: SxProps }) {
   const iconBgColor =
@@ -42,36 +30,26 @@ export function TaskIcon({ type, sx }: { type: TaskType; sx?: SxProps }) {
     } ?? '#9A53FF';
 
   const iconComponent = useMemo(() => {
-    switch (type) {
-      case 'self_verify':
-        return InsertLinkIcon;
-      case 'quiz':
-        return QuizIcon;
-      case 'token_hold':
-        return MonetizationOnIcon;
-      case 'nft_hold':
-        return PhotoCameraBack;
-      case 'meeting_code':
-        return NumbersIcon;
-      case 'twitter_follow':
-        return Twitter;
-      case 'twitter_retweet':
-        return Twitter;
-      case 'twitter_tweet':
-        return Twitter;
-      case 'github_contribute':
-        return GitHubIcon;
-      case 'github_prs':
-        return GitHubIcon;
-      case 'snapshot':
-        return ElectricBoltIcon;
-      default:
-        return null;
-    }
+    const types = {
+      self_verify: InsertLinkIcon,
+      quiz: QuizIcon,
+      token_hold: MonetizationOnIcon,
+      nft_hold: PhotoCameraBack,
+      meeting_code: NumbersIcon,
+      twitter_follow: Twitter,
+      twitter_retweet: Twitter,
+      twitter_tweet: Twitter,
+      github_contribute: GitHubIcon,
+      github_prs: GitHubIcon,
+      snapshot: ElectricBoltIcon,
+    };
+
+    return types[type] || null;
   }, [type]);
 
   return (
     <Box
+      display="flex"
       component="span"
       bgcolor={iconBgColor[type]}
       padding={1.5}
