+++ apps/website/components/atoms/task-icon/types.ts
@@ -1,0 +1,12 @@
+export type TaskType =
+  | 'self_verify'
+  | 'quiz'
+  | 'token_hold'
+  | 'nft_hold'
+  | 'meeting_code'
+  | 'twitter_follow'
+  | 'twitter_retweet'
+  | 'twitter_tweet'
+  | 'github_contribute'
+  | 'github_prs'
+  | 'snapshot';
